# docs/swagger.yaml
basePath: /
definitions:
...
  commons.Pagination:
    properties:
      page_number:
        type: integer
      page_size:
        type: integer
      total_item:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.CreateQuoteRequest:
    properties:
      customer_id:
        type: string
      discount:
        type: number
      items:
        items:
          $ref: '#/definitions/valueobject.QuoteItem'
        type: array
      notes:
        type: string
      quote_date:
        type: integer
      valid_until:
        type: integer
    type: object
...
  models.Customer:
    properties:
      address:
        type: string
      country:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/valueobject.Metadata'
      name:
        type: string
      phone:
        type: string
      picture_url:
        type: string
      postal_code:
        type: string
      social_name:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.Quote:
    properties:
      customer:
        $ref: '#/definitions/models.Customer'
      discount:
        type: number
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/valueobject.QuoteItem'
        type: array
      metadata:
        $ref: '#/definitions/valueobject.Metadata'
      notes:
        type: string
      quote_date:
        type: integer
      status:
        $ref: '#/definitions/valueobject.QuoteStatus'
      valid_until:
        type: integer
    type: object
  models.Quotes:
    properties:
      pagination:
        $ref: '#/definitions/commons.Pagination'
      quotes:
        items:
          $ref: '#/definitions/models.Quote'
        type: array
    type: object
...
  models.UpdateQuoteRequest:
    properties:
      customer_id:
        type: string
      discount:
        type: number
      items:
        items:
          $ref: '#/definitions/valueobject.QuoteItem'
        type: array
      notes:
        type: string
      quote_date:
        type: integer
      status:
        $ref: '#/definitions/valueobject.QuoteStatus'
      valid_until:
        type: integer
    type: object
...
  valueobject.Metadata:
    properties:
      created_at:
        type: integer
      created_by:
        type: string
      deleted_at:
        type: integer
      deleted_by:
        type: string
      updated_at:
        type: integer
      updated_by:
        type: string
    type: object
...
  valueobject.QuoteItem:
    properties:
      description:
        type: string
      quantity:
        type: integer
      tax_rate:
        type: number
      unit_price:
        type: number
    type: object
  valueobject.QuoteStatus:
    enum:
    - QUOTE_STATUS_UNSPECIFIED
    - PENDING
    - APPROVED
    - REJECTED
    type: string
    x-enum-varnames:
    - QuoteStatusUnspecified
    - QuoteStatusPending
    - QuoteStatusApproved
    - QuoteStatusRejected
host: localhost:9000
info:
  contact:
    email: orkys.com@gmail.com
    name: API Support
    url: http://www.example.com/support
  description: This is the API documentation for the Orkys API Gateway
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: Orkys API
  version: "1.0"
paths:
...
  /quote:
    get:
      consumes:
      - application/json
      description: Returns a list of quotes with pagination.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of events per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: The quote object list, and pagination details
          schema:
            $ref: '#/definitions/models.Quotes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Returns a list of quotes.
      tags:
      - quote
    post:
      consumes:
      - application/json
      description: Creates a new quote with the provided details.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: The quote object
        in: body
        name: quote
        required: true
        schema:
          $ref: '#/definitions/models.CreateQuoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: The created quote object
          schema:
            $ref: '#/definitions/models.Quote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Creates a new quote.
      tags:
      - quote
  /quote/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a quote by ID.
      parameters:
      - description: Quote ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Deletes a quote.
      tags:
      - quote
    get:
      consumes:
      - application/json
      description: Returns a quote by ID with all the details.
      parameters:
      - description: Quote ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The quote object
          schema:
            $ref: '#/definitions/models.Quote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Returns a quote by ID.
      tags:
      - quote
    patch:
      consumes:
      - application/json
      description: Updates a quote with the provided details.
      parameters:
      - description: Quote ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: The quote object
        in: body
        name: quote
        required: true
        schema:
          $ref: '#/definitions/models.UpdateQuoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: The updated quote object
          schema:
            $ref: '#/definitions/models.Quote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Updates a quote.
      tags:
      - quote
...
swagger: "2.0"